# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Oshkelosh",
    "description": "# Oshkelosh API\n\nModular e-commerce backend: core REST API plus **pluggable addons** for suppliers,\npayments, notifications, and storefront frontends.\n\n## Documentation (repository paths)\n\n| Topic | Path |\n|-------|------|\n| Documentation index | `docs/README.md` |\n| All plugins (overview) | `app/addons/README.md` |\n| Storefront / SPA development | `app/addons/frontends/README.md` |\n| Payment processors | `app/addons/payments/README.md` |\n| Suppliers / fulfillment | `app/addons/suppliers/README.md` |\n| Email / notifications | `app/addons/notifications/README.md` |\n| OpenAPI reference | `docs/api/OPENAPI.md` |\n\n## Authentication\n\nMost customer endpoints require:\n\n```\nAuthorization: Bearer <access_token>\n```\n\nObtain tokens via `POST /api/v1/auth/login` or `POST /api/v1/auth/register`.\n\nAdmin JSON endpoints require a user with `is_admin=true`.\n\n## Storefront SPA contract\n\n1. Load branding: `GET /api/v1/storefront/config`\n2. Optional early CSS: `GET /api/v1/storefront/theme.css`\n3. Use core APIs for catalog, cart, orders, checkout\n\n## Addon routes\n\nAddon paths follow:\n\n- **API:** `/api/v1/{category}s/{addon_id}/...` (example: `/api/v1/payments/stripe/checkout`)\n- **Admin HTML:** `/admin/{category}s/{addon_id}/...` (configure credentials and options)\n\nEnable and configure addons in the admin panel at `/admin/addons`.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/orders": {
      "post": {
        "tags": [
          "orders",
          "orders"
        ],
        "summary": "Create order from cart",
        "description": "Create a new order from the current user's cart. Inventory is reserved when payment is confirmed (status paid).",
        "operationId": "create_order_api_v1_orders_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "orders",
          "orders"
        ],
        "summary": "List user's orders",
        "description": "Return a paginated list of the current user's orders.",
        "operationId": "list_orders_api_v1_orders_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Items per page"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by order status",
              "title": "Status"
            },
            "description": "Filter by order status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Orders Api V1 Orders Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}/checkout": {
      "post": {
        "tags": [
          "orders",
          "orders"
        ],
        "summary": "Create payment checkout session",
        "description": "Start checkout with the enabled payment processor for a pending order.",
        "operationId": "checkout_order_api_v1_orders__order_id__checkout_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Checkout Order Api V1 Orders  Order Id  Checkout Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}": {
      "get": {
        "tags": [
          "orders",
          "orders"
        ],
        "summary": "Get order detail",
        "description": "Return the full detail of a single order belonging to the current user.",
        "operationId": "get_order_api_v1_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}/cancel": {
      "post": {
        "tags": [
          "orders",
          "orders"
        ],
        "summary": "Cancel order",
        "description": "Cancel an order if its status allows cancellation (pending, paid, shipped).",
        "operationId": "cancel_order_api_v1_orders__order_id__cancel_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/admin": {
      "get": {
        "tags": [
          "orders",
          "orders"
        ],
        "summary": "List all orders (admin)",
        "description": "Return a paginated list of all orders. Admin only.",
        "operationId": "admin_list_orders_api_v1_orders_admin_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Items per page"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by order status",
              "title": "Status"
            },
            "description": "Filter by order status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Admin List Orders Api V1 Orders Admin Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/admin/{order_id}/status": {
      "patch": {
        "tags": [
          "orders",
          "orders"
        ],
        "summary": "Update order status (admin)",
        "description": "Update the status of an order. Admin only.",
        "operationId": "admin_update_order_status_api_v1_orders_admin__order_id__status_patch",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderUpdateStatus"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/orders": {
      "get": {
        "tags": [
          "admin",
          "admin"
        ],
        "summary": "Admin List Orders",
        "description": "List all orders (admin).",
        "operationId": "admin_list_orders_api_v1_admin_orders_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_OrderRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/": {
      "get": {
        "summary": "Admin Dashboard",
        "description": "Admin dashboard with key metrics.",
        "operationId": "admin_dashboard_admin__get",
        "parameters": [
          {
            "name": "oshkelosh_admin",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oshkelosh Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/orders": {
      "get": {
        "summary": "Admin Orders List",
        "description": "List orders with pagination and optional status filter.",
        "operationId": "admin_orders_list_admin_orders_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "name": "oshkelosh_admin",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oshkelosh Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/orders/{order_id}": {
      "get": {
        "summary": "Admin Order Detail",
        "description": "Show a single order with its line items.",
        "operationId": "admin_order_detail_admin_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          },
          {
            "name": "oshkelosh_admin",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oshkelosh Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderItemRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "order_id": {
            "type": "integer",
            "title": "Order Id"
          },
          "product_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "product_sku": {
            "type": "string",
            "title": "Product Sku"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit_price_cents": {
            "type": "integer",
            "title": "Unit Price Cents"
          },
          "unit_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Unit Price",
            "default": "0.00"
          },
          "total_price_cents": {
            "type": "integer",
            "title": "Total Price Cents"
          },
          "total_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Price",
            "default": "0.00"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_id",
          "product_id",
          "product_name",
          "product_sku",
          "quantity",
          "unit_price_cents",
          "total_price_cents",
          "created_at"
        ],
        "title": "OrderItemRead"
      },
      "OrderRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "total_cents": {
            "type": "integer",
            "title": "Total Cents"
          },
          "total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total",
            "default": "0.00"
          },
          "tax_cents": {
            "type": "integer",
            "title": "Tax Cents"
          },
          "tax": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Tax",
            "default": "0.00"
          },
          "shipping_cents": {
            "type": "integer",
            "title": "Shipping Cents"
          },
          "shipping": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Shipping",
            "default": "0.00"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "shipping_address": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipping Address"
          },
          "billing_address": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Address"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemRead"
            },
            "type": "array",
            "title": "Items"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "user_id",
          "status",
          "total_cents",
          "tax_cents",
          "shipping_cents",
          "currency",
          "shipping_address",
          "billing_address",
          "notes",
          "created_at",
          "updated_at"
        ],
        "title": "OrderRead"
      },
      "OrderUpdateStatus": {
        "properties": {
          "status": {
            "type": "string",
            "pattern": "^(pending|paid|shipped|delivered|cancelled)$",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "OrderUpdateStatus"
      },
...
      "PaginatedResponse_OrderRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "PaginatedResponse[OrderRead]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}